+++ apps/fret-examples/src/shaded_demo.rs
@@ -38,7 +38,7 @@ impl ShadedDemoDriver {
         let mut lower1: Vec<DataPoint> = Vec::with_capacity(n);
 
         for i in 0..n {
-            let t = i as f32 / (n - 1) as f32;
+            let t = i as f64 / (n - 1) as f64;
             let x = t * 10.0;
 
             let base0 = (x * 1.05).sin() * 0.75 + (x * 0.25).cos() * 0.10;
